pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nss-host"
version = "0.1.0"
description = "Raspberry Pi 5 Host Driver for NRWA-T6 Reaction Wheel Emulator"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["reaction-wheel", "nsp", "rs485", "raspberry-pi", "aerospace"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
    "Topic :: System :: Hardware",
]

dependencies = [
    "pyserial>=3.5",
    "gpiozero>=2.0",
    "textual>=0.40.0",
    "pydantic>=2.0",
    "rich>=13.0",
    "tomli>=2.0; python_version<'3.11'",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "pytest-timeout>=2.1",
    "black>=23.0",
    "ruff>=0.1.0",
    "mypy>=1.5",
]
plots = [
    "numpy>=1.24",
    "matplotlib>=3.7",
]
full = [
    "nss-host[dev,plots]",
]

[project.scripts]
nss-tui = "nss_host.tui.tui:main"
nss-send = "tools.nsp_send:main"
nss-dump = "tools.dump_telemetry:main"
nss-bench = "tools.bench_rtt:main"
nss-fuzz = "tools.fuzz_cmds:main"
nss-record = "tools.record:main"

[project.urls]
Homepage = "https://github.com/yourusername/nsrw_test"
Repository = "https://github.com/yourusername/nsrw_test"
Issues = "https://github.com/yourusername/nsrw_test/issues"

[tool.setuptools]
packages = ["nss_host", "nss_host.tui", "nss_host.scenarios", "nss_host.tests", "tools"]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers"
testpaths = ["nss_host/tests"]
markers = [
    "unit: Unit tests (no hardware required)",
    "hil: Hardware-in-the-loop tests (requires connected emulator)",
    "slow: Slow-running tests",
]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "W", "I", "N", "UP"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
